requirements.txt
google-cloud-vision
gspread
oauth2client
PyYAML
python-dotenv